#16 Design and implement bills grid
* Refactored sample data initialization class and provided more sample data
* Changed cost type in bill item model
* Removed mock data provider, controller and server context path configuration
* Bumped up version on vaadin dependency
* Implemented simple grid to display bills

+++ backend/src/main/java/com/pwawrzyniak/tlog/backend/model/BillItem.java
@@ -16,6 +16,7 @@ import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 import java.util.Set;
 
 @Data
@@ -32,7 +33,7 @@ public class BillItem {
   private Long id;
 
   @Column(nullable = false)
-  private Long cost;
+  private BigDecimal cost;
 
   @Column
   private String description;
